requirements.txt
git+https://github.com/pool-energy/chia-blockchain.git@release/2.5.5#egg=chia-blockchain
git+https://github.com/pool-energy/pool.git@main#egg=pool
chia_rs==0.27.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
requests==2.32.5
django-filter==25.1
psycopg2==2.9.10
channels==4.3.1
cachetools==6.2.0
qrcode==8.2
pytz==2025.2
influxdb-client==1.49.0
py-spy==0.4.1
uvicorn==0.35.0
gunicorn==23.0.0
websockets==15.0.1
